# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SweetPOS API for Malaysian candy and cake shop
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Orders
...
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get sales summary (today, this week, this month)
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [today, week, month]
      responses:
        "200":
          description: Sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /reports/top-products:
    get:
      operationId: getTopProducts
      tags: [reports]
      summary: Get top selling products
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [today, week, month]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /reports/sales-by-day:
    get:
      operationId: getSalesByDay
      tags: [reports]
      summary: Get daily sales for a period (chart data)
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Sales by day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesByDay"
  /reports/sales-by-category:
    get:
      operationId: getSalesByCategory
      tags: [reports]
      summary: Get sales breakdown by category
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [today, week, month]
      responses:
        "200":
          description: Sales by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesByCategory"
  /reports/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [reports]
      summary: Get recent orders for activity feed
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
          description: Price in MYR
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        stock:
          type: integer
          description: Current stock quantity
        taxable:
          type: boolean
          description: Whether 8% SST applies
        active:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - price
        - stock
        - taxable
        - active
        - createdAt
        - updatedAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        categoryId:
          type: ["integer", "null"]
        sku:
          type: ["string", "null"]
        stock:
          type: integer
        taxable:
          type: boolean
        active:
          type: boolean
        imageUrl:
          type: ["string", "null"]
      required:
        - name
        - price
        - stock
        - taxable
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        categoryId:
          type: ["integer", "null"]
        sku:
          type: ["string", "null"]
        stock:
          type: integer
        taxable:
          type: boolean
        active:
          type: boolean
        imageUrl:
          type: ["string", "null"]

    # Order
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        taxRate:
          type: number
          description: Tax rate applied (e.g. 0.08 for 8%)
        subtotal:
          type: number
        tax:
          type: number
        total:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - unitPrice
        - taxRate
        - subtotal
        - tax
        - total
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
          enum: [pending, completed, cancelled, refunded]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        taxTotal:
          type: number
        total:
          type: number
        paymentMethod:
          type: ["string", "null"]
          enum: ["cash", "card", "ewallet", null]
        amountPaid:
          type: ["number", "null"]
        change:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        staffId:
          type: ["integer", "null"]
        staffName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderNumber
        - status
        - items
        - subtotal
        - taxTotal
        - total
        - createdAt
        - updatedAt
...
    ReportSummary:
      type: object
      properties:
        totalSales:
          type: number
        totalOrders:
          type: integer
        totalTax:
          type: number
        averageOrderValue:
          type: number
        period:
          type: string
      required:
        - totalSales
        - totalOrders
        - totalTax
        - averageOrderValue
        - period
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        categoryName:
          type: ["string", "null"]
        quantitySold:
          type: integer
        revenue:
          type: number
      required:
        - productId
        - productName
        - quantitySold
        - revenue
    SalesByDay:
      type: object
      properties:
        date:
          type: string
        totalSales:
          type: number
        totalOrders:
          type: integer
      required:
        - date
        - totalSales
        - totalOrders
    SalesByCategory:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: string
        totalSales:
          type: number
        percentage:
          type: number
      required:
        - categoryName
        - totalSales
        - percentage