# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SweetPOS API for Malaysian candy and cake shop
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  sku?: string | null;
  stock: number;
  taxable: boolean;
  active?: boolean;
  /** @nullable */
  imageUrl?: string | null;
}