# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SweetPOS API for Malaysian candy and cake shop
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** Price in MYR */
  price: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  sku?: string | null;
  /** Current stock quantity */
  stock: number;
  /** Whether 8% SST applies */
  taxable: boolean;
  active: boolean;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}