# lib/api-zod/src/generated/types/salesByCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SweetPOS API for Malaysian candy and cake shop
 * OpenAPI spec version: 0.1.0
 */

export interface SalesByCategory {
  /** @nullable */
  categoryId?: number | null;
  categoryName: string;
  totalSales: number;
  percentage: number;
}